Fix date/datetime/bit encode error (#200)
* Enhance dbg invoke and add dag as schemaful function

* Add basic sql parse to dag

* Column id starts from 1

* Fix value to ref

* Add basic dag test

* Fix dag bugs and pass 1st mock test

* Make dag go normal routine and add mock dag

* Add todo

* Add comment

* Fix gcc compile error

* Enhance dag test

* Address comments

* Enhance mock sql -> dag compiler and add project test

* Mock sql dag compiler support more expression types and add filter test

* Add topn and limit test

* Add agg for sql -> dag parser and agg test

* Add dag specific codec

* type

* Update codec accordingly

* Remove cop-test

* Pass tests after merging master

* Copy some changes from xufei

* Enable date/datetime test

* Enable date/datetime test

* Refine code

* Adjust date/datetime tiflash rep to UInt

* Fix datetime to Int

* Typo

+++ dbms/src/Storages/Transaction/Datum.cpp
@@ -44,7 +44,7 @@ struct DatumOp<tp, typename std::enable_if<tp == TypeDate || tp == TypeDatetime 
         if constexpr (tp == TypeDate)
         {
             auto date = date_lut.makeDayNum(year, month, day);
-            copy = static_cast<Int64>(date);
+            copy = static_cast<UInt64>(date);
         }
         else
         {
@@ -86,7 +86,7 @@ struct DatumOp<tp, typename std::enable_if<tp == TypeDate || tp == TypeDatetime 
         }
         else
         {
-            time_t date_time(static_cast<UInt64>(orig.get<UInt64>()));
+            time_t date_time(orig.get<Int64>());
             values = date_lut.getValues(date_time);
             hour = date_lut.toHour(date_time);
             minute = date_lut.toMinute(date_time);
@@ -100,7 +100,7 @@ struct DatumOp<tp, typename std::enable_if<tp == TypeDate || tp == TypeDatetime 
     static bool overflow(const Field &, const ColumnInfo &) { return false; }
 };
 
-/// Specialized for integer types less than 64b, checks overflow.
+/// Specialized for integer types less than 64 bit, checks overflow.
 template <TP tp>
 struct DatumOp<tp, typename std::enable_if<tp == TypeTiny || tp == TypeShort || tp == TypeLong || tp == TypeInt24>::type>
 {
